+++ apps/hops-bestilling/src/test/kotlin/no/nav/helse/hops/componentTests/ValidationErrorTest.kt
@@ -6,17 +6,12 @@ import kotlinx.coroutines.runBlocking
 import no.nav.helse.hops.domain.isAllOk
 import no.nav.helse.hops.domain.toJson
 import no.nav.helse.hops.infrastructure.FhirResourceValidatorHapi
-import no.nav.helse.hops.infrastructure.KafkaFhirResourceSerializer
 import no.nav.helse.hops.infrastructure.KoinBootstrapper
+import no.nav.helse.hops.testUtils.KafkaMock
 import no.nav.helse.hops.testUtils.ResourceLoader
+import no.nav.helse.hops.testUtils.addFhirMessage
 import org.apache.kafka.clients.consumer.Consumer
-import org.apache.kafka.clients.consumer.ConsumerRecord
-import org.apache.kafka.clients.consumer.MockConsumer
-import org.apache.kafka.clients.consumer.OffsetResetStrategy
-import org.apache.kafka.clients.producer.MockProducer
 import org.apache.kafka.clients.producer.Producer
-import org.apache.kafka.common.TopicPartition
-import org.apache.kafka.common.serialization.Serializer
 import org.hl7.fhir.instance.model.api.IBaseResource
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.MessageHeader
@@ -27,13 +22,9 @@ import org.koin.dsl.module
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-private const val TOPIC = "helseopplysninger.bestilling"
-private const val PARTITION = 0
-
-class ValidationErrorTest : java.io.Closeable {
-    private val consumerMock = createMockConsumer()
-    private val producerMock =
-        MockProducer(true, Serializer<Unit> { _, _ -> ByteArray(0) }, KafkaFhirResourceSerializer())
+class ValidationErrorTest {
+    private val consumerMock = KafkaMock.createConsumer()
+    private val producerMock = KafkaMock.createProducer()
     private val koinApp = startKoin {
         val testKoinModule = module(override = true) {
             single<Producer<Unit, IBaseResource>> { producerMock }
@@ -47,7 +38,7 @@ class ValidationErrorTest : java.io.Closeable {
     @Test
     fun `invalid message should result in response with operation-outcome`() {
         val requestMessage = ResourceLoader.asFhirResource<Bundle>("/fhir/invalid-message-warning-on-name.json")
-        consumerMock.schedulePollTask { consumerMock.addRecord(ConsumerRecord(TOPIC, PARTITION, 0, Unit, requestMessage)) }
+        consumerMock.addFhirMessage(requestMessage)
 
         koinApp.close()
 
@@ -73,24 +64,5 @@ class ValidationErrorTest : java.io.Closeable {
             val outcome = FhirResourceValidatorHapi.validate(responseMessage)
             assertTrue(outcome.isAllOk(), outcome.toJson())
         }
-    }
-
-    override fun close() {
-        consumerMock.close()
-        producerMock.close()
-    }
-}
-
-private fun createMockConsumer(): MockConsumer<Unit, IBaseResource> {
-    return MockConsumer<Unit, IBaseResource>(OffsetResetStrategy.EARLIEST).apply {
-        schedulePollTask {
-            rebalance(listOf(TopicPartition(TOPIC, 0)))
-        }
-
-        val startOffsets = HashMap<TopicPartition, Long>()
-        val tp = TopicPartition(TOPIC, PARTITION)
-        startOffsets[tp] = 0L
-
-        updateBeginningOffsets(startOffsets)
     }
 }
